pyproject.toml
[tool.poetry]
name = "django-assignment"
version = "0.1.0"
description = ""
authors = ["maskudo <manujkarki101@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
django-stubs = {extras = ["compatible-mypy"], version = "^4.2.6"}
djangorestframework = "^3.14.0"
reportlab = "^4.0.7"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "myproject.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
